Run rally only when opted with VERIFY_INSTALL env var (#1282)
* Currently defaults to False

+++ bootstrap/shared/shared_configure_chef.sh
@@ -177,16 +177,18 @@ else
     do_on_node "$vm" "sudo chef-client"
   done
 
-  # Do rally setup at the end without polluting the main run lists
-  # also this will make sure the cluster is fully provisioned and ready for sanity checks
-  do_on_node vm-bootstrap "sudo chef-client -o bcpc::rally,bcpc::rally-deployments"
-  if [ -z "$RALLY_SCENARIOS_DIR" ]; then
-    echo "Not copying RALLY scenario files, set RALLY_SCENARIOS_DIR pointing to your custom rally scenarios repo"
-  else
-    OPTIONS=$(vagrant ssh-config vm-bootstrap | awk -v ORS=' ' 'NF && !/Host / {print "-o " $1 "=" $2}')
-    # Remove / at the end
-    RALLY_SCENARIOS_DIR="${RALLY_SCENARIOS_DIR%/}"
-    rsync -avz --delete --exclude=.git -e "ssh $OPTIONS" "$RALLY_SCENARIOS_DIR" vm-bootstrap:rally
-    do_on_node vm-bootstrap "sudo chef-client -o bcpc::rally-run"
+  if [ "$VERIFY_CLUSTER" = "1" ]; then
+    # Do rally setup at the end without polluting the main run lists
+    # also this will make sure the cluster is fully provisioned and ready for sanity checks
+    do_on_node vm-bootstrap "sudo chef-client -o bcpc::rally,bcpc::rally-deployments"
+    if [ -z "$RALLY_SCENARIOS_DIR" ]; then
+      echo "Not copying RALLY scenario files, set RALLY_SCENARIOS_DIR pointing to your custom rally scenarios repo"
+    else
+      OPTIONS=$(vagrant ssh-config vm-bootstrap | awk -v ORS=' ' 'NF && !/Host / {print "-o " $1 "=" $2}')
+      # Remove / at the end
+      RALLY_SCENARIOS_DIR="${RALLY_SCENARIOS_DIR%/}"
+      rsync -avz --delete --exclude=.git -e "ssh $OPTIONS" "$RALLY_SCENARIOS_DIR" vm-bootstrap:rally
+      do_on_node vm-bootstrap "sudo chef-client -o bcpc::rally-run"
+    fi
   fi
 fi
